Remove rule that checks if HTTP Range header begins with 'bytes=0-'
The current implementation states that 'Range Header exists and begins with 0 - normal browsers don't do this.'. This has changed and can now be seen in both Firefox (tested with 26.0 and 27.0) and Chromium (tested with 33.0.1750.149). It only happens when requesting continuous content which requires loading, e.g. a video.

An example to reproduce: Go to `http://www.videojs.com/`, open development tools in Firefox/Chromium, go to network, klick on the video on the website and watch the requests for the video stream. It will contain: `Range: bytes=0-`.

Signed-off-by: Dominik Richter <dominik.richter@gmail.com>